feat(game): Implement dynamic and combat-based item durability loss
This commit introduces a durability loss system for equipped armor, applying wear and tear during combat encounters.

- Armor now has a chance to lose durability during both PvE and PvP combat, configurable via `PvEDurabilityLossRate` and `PvPDurabilityLossRate` in `World.json`.
- Refactors death-based durability loss, now referencing `DeathDurabilityLossRatio` from item configurations.
- Ensures destroyed items (those with zero durability) no longer contribute to character stats or apply any active buffs, correctly removing their benefits until repaired.

+++ AAEmu.Game/Models/Game/Items/ItemConfig.cs
@@ -7,7 +7,7 @@ public sealed class ItemConfig
     public float DurabilityConst { get; set; }
     public float HoldableDurabilityConst { get; set; }
     public float WearableDurabilityConst { get; set; }
-    public int DeathDurabilityLossRatio { get; set; }
+    public byte DeathDurabilityLossRatio { get; set; }
     public int ItemStatConst { get; set; }
     public int HoldableStatConst { get; set; }
     public int WearableStatConst { get; set; }
